backend/requirements.txt
# requirements.txt
# ================================================
# KI-Protokoll-Stack 2025 – ws-ki (Projektstand 2025-07-01)
# Python-Version: >=3.11,<3.12
# Empfohlene Nutzung: venv + pyproject.toml + requirements.txt
# ================================================

# Python-Version für das Projekt
# (zur Kontrolle, eigentliche Enforcement via pyproject.toml)
# requires-python = ">=3.11,<3.12"

# === BASIS: FastAPI, ClickHouse, Docker ===
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.13
h11<0.15,>=0.13
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
clickhouse-connect==0.7.9

# === INDICATORS: DataFrame, Math, TA, Signalprocessing ===
pandas==2.2.2
numpy==1.26.4
pandas-ta==0.3.14b0        # Für ALMA, TA-Indikatoren
scipy==1.13.1              # Für FFT, Signalverarbeitung (Spectral/Extrema)
plotly==5.22.0             # Für optionale Plot-Ausgabe der Elliott Waves

# === Empfohlene Erweiterungen für Web, Requests, EVM/Krypto ===
httpx==0.26.0              # Async HTTP-Client (REST, Webhook, API)
web3==6.19.0               # EVM-Integration, Whale-Detection, Ethereum/WebSocket

# === TRADING SYSTEM: Grid Trading, Risk Management ===
redis==5.0.0               # Redis für Caching und Session Management
kafka-python==2.0.2        # Kafka für Messaging zwischen Services
aioredis==2.0.1           # Async Redis Client
aiokafka==0.12.0          # Async Kafka Client für bessere Performance
celery==5.3.4             # Background Task Processing
kombu==5.3.4              # Celery Message Transport

# === (Optional: Bilder, AI etc.) ===
pillow==10.3.0
# openai==1.30.1            # AI-Integration, falls benötigt

# === (Optional: Benachrichtigungen, Scheduler, Telegram etc.) ===
python-telegram-bot==20.8
aiojobs==1.2.1

# === (Optional: Entwicklungs-/Analyse-Tools) ===
ipython==8.25.0
jupyterlab==4.2.2

# === (Optional: Typing, Linting, Tests) ===
mypy==1.10.0
pytest==8.2.2
